requirements.txt
pillow
pytesseract
deep-translator
pymupdf
streamlit
